Move MessageStorageHandler to its own module
